requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aisuite==0.1.6
altair==5.4.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.23
botocore==1.35.23
bottle==0.13.1
bottle-websocket==0.2.9
Brotli==1.1.0
bson==0.5.10
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clipman==3.3.1
cloudscraper==1.2.71
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.7
contourpy==1.2.1
cryptography==44.0.0
ctransformers==0.2.27
curl_cffi==0.7.4
cycler==0.12.1
dataclasses-json==0.6.7
dbus-next==0.2.3
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.30.0
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
Eel==0.17.0
emoji==2.14.0
executing==2.1.0
fastapi==0.112.2
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
g4f==0.3.8.1
gast==0.6.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gradio_client==1.5.0
graphviz==0.20.3
greenlet==3.0.3
groq==0.9.0
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.66.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.5
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaitaistruct==0.10
keras==3.5.0
kiwisolver==1.4.5
langchain==0.3.7
langchain-aws==0.2.1
langchain-community==0.2.14
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.22
langgraph-checkpoint==1.0.10
langsmith==0.1.125
libclang==18.1.1
llama-cloud==0.0.17
llama-index==0.11.10
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.10
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.9
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-utils-workflow==0.2.1
llama-parse==0.5.6
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.2.0
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.4.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
openai==1.45.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipdeptree==2.23.3
platformdirs==4.2.2
playsound==1.2.2
plotly==5.23.0
portalocker==2.10.1
primp==0.8.1
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.0
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydivert==2.1.0
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyOpenSSL==24.3.0
pyparsing==3.1.2
pypdf==4.3.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
pyreqwest_impersonate==0.5.3
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.1
pyvis==0.3.2
pywhatkit==5.4
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.11.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.27.1
selenium-wire==5.1.0
sentence-transformers==3.0.1
setuptools==74.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.17.0
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tls-client==1.0.1
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.4
waitress==3.0.0
watchdog==4.0.2
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
webscout==6.4
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
whichcraft==0.6.1
wikipedia==1.4.0
winotify==1.1.0
WMI==1.5.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yaspin==3.1.0
zipp==3.20.0
zope.event==5.0
zope.interface==7.0.3
zstandard==0.23.0
